// src/api/data_anno_task.ts
export type TaskInfo = {
    anno_project_id: string;
    member_user_id: string;
    resource_id: string;
    content: string;
    store_as_file: boolean;
    file_ext: string;
    done: boolean;
};